Added two-phase training for ConvModel, adapted perf_test_cam
Now the conv model can train first the behavioral cloning part, and then the AE part, so that the embedded part of BC is fixed. To save time, it caches the hidden outputs of BC. 
perf_test_cam here mixes the test set with some training examples, to check if their error is lower.